refactor: rename and move time controller for consistency
