+++ packages/calcite-components/src/controllers/useTime.ts
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { PropertyValues } from "lit";
 import { GenericController, T9nMeta, toFunction } from "@arcgis/lumina/controllers";
 import { GenericT9nStrings } from "@arcgis/components-utils";
@@ -21,12 +22,12 @@ import {
   MinuteOrSecond,
   parseTimeString,
   toISOTimeString,
-} from "../../utils/time";
-import { decimalPlaces, getDecimals } from "../../utils/math";
-import { isValidNumber } from "../../utils/number";
-import { capitalizeWord } from "../../utils/text";
-import { NumberingSystem, SupportedLocale } from "../../utils/locale";
-import { numberKeys } from "../../utils/key";
+} from "../utils/time";
+import { decimalPlaces, getDecimals } from "../utils/math";
+import { isValidNumber } from "../utils/number";
+import { capitalizeWord } from "../utils/text";
+import { NumberingSystem, SupportedLocale } from "../utils/locale";
+import { numberKeys } from "../utils/key";
 
 export interface TimeComponent extends LitElement {
   /**
@@ -141,7 +142,7 @@ type TimeProperties = {
   second: string;
 };
 
-export class TimeController extends GenericController<TimeProperties, TimeComponent> {
+class TimeController extends GenericController<TimeProperties, TimeComponent> {
   //#region Private Properties
 
   fractionalSecond: string;
